Refactored OpenAI API calls to backend. Filtered duplicated responses when reading from Google Sheet.
