+++ src/app/api/prompt/openai.js
@@ -1,0 +1,8 @@
+import OpenAI from 'openai';
+
+const openai = new OpenAI({
+    apiKey: process.env.OPEN_API_KEY,
+    dangerouslyAllowBrowser: true
+});
+
+export default openai;
